api/pyproject.toml
[project]
name = "yourharmony-api"
version = "0.1.0"
description = "RESTful API for YourHarmony educational platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.*",
    "uvicorn[standard]==0.32.*",
    "sqlalchemy[asyncio]==2.0.*",
    "asyncpg==0.30.*",
    "pydantic==2.*",
    "pydantic-settings==2.*",
    "python-jose[cryptography]==3.3.*",
    "passlib[bcrypt]==1.7.*",
    "python-multipart==0.0.*",
]

[dependency-groups]
dev = [
    "pytest==8.*",
    "pytest-asyncio==0.24.*",
    "httpx==0.28.*",
    "anyio[trio]",
    "pytest-cov",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
